increase granularity of hide by using moderation.isUserHidden
instead of tracking hide state per channel context in cable-client, we
call cable-core's instance of ModerationSystem (which is always kept up
to date).

why we rely on cable-core's ModerationSystem instance instead of e.g.
the moderation event sent from cable-core when a new hide/unhide action
happens is because we'd have to implement additional logic for
determining *if* the event should be applied to our state or not (maybe
the author of the moderation action was not a moderation authority - we
can still display taken actions even if they have no effect for us as a
way to 'discover' who we are aligned with in terms of moderation
decisions). doing it this way we also don't duplicate state management
outside of the library taking are of it.

in the current solution in cable-client, we wait until updating
moderation state until the `moderation/action-update` even has fired,
which is when we iterate over all the users and update their `hidden
attribute`. what we could do instead of iterate over all the users, is
use the e.g. moderation/hide event from core, which passes us the
recipient public key, and then use that to query the moderation state
& update if changed. for later though, as it may have some pitfalls in
terms of when the ModerationSystem instance is updated and when the
single event is fired