refine updating a user's moderation state
this commit replaces the previous coarse method of iterating over all
users. we now use the passed-in post - the post which caused the
moderation state to change, as a result of a moderation authority
authoring a post. using the type of post, the recipients field of the
post, and the type of action, we update the state of the targeted
recipients by querying core's ModerationSystem instance.